refactor: fixed inconsistent naming gf-557
