+++ apps/frontend/src/libs/hooks/hooks.ts
@@ -2,7 +2,7 @@ export { useAppDispatch } from "./use-app-dispatch/use-app-dispatch.hook.js";
 export { useAppForm } from "./use-app-form/use-app-form.hook.js";
 export { useAppSelector } from "./use-app-selector/use-app-selector.hook.js";
 export { useHandleClickOutside } from "./use-handle-click-outside/use-handle-click-outside.hook.js";
-export { useInfiniteScroll } from "./use-infinity-scroll/use-infinity-scroll.hook.js";
+export { useInfiniteScroll } from "./use-infinite-scroll/use-infinite-scroll.hook.js";
 export { useIntersectionObserver } from "./use-intersection-observer/use-intersection-observer.hook.js";
 export { useModal } from "./use-modal-state/use-modal-state.hook.js";
 export { usePagination } from "./use-pagination/use-pagination.hook.js";
